% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs.couples}
\alias{acs.couples}
\title{Married Couples from ACS 2011-2015}
\format{A data frame containing data on 1000 married couples from the 2011-2015 ACS data drawn randomly from
       the states of New York and California. The data frame contains the following variables for each couple:
       \describe{
         \item{state}{State of residence. This can be used to select appropriate cluster to sample alternate spouses}
         \item{hhwt}{Household weight for the given couple}
         \item{idw}{unique id of wife}
         \item{agew}{age of wife}
         \item{racew}{race of wife. Either "W" (white) or "B" (black)}
         \item{idh}{unique id of husband}
         \item{ageh}{age of husband}
         \item{raceh}{race of husband. Either "W" (white) or "B" (black)}
       }}
\source{
\url{https://usa.ipums.org/usa/}
}
\usage{
acs.couples
}
\description{
A dataset containing information on a sample of married couples from the American Community Survey 2011-2015
data for the states of New York and California. Data were downloaded from IPUMS.
}
\keyword{datasets}
