% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs.malealters}
\alias{acs.malealters}
\title{Alternate male partners from ACS 2011-2015}
\format{A data frame containing data on 1000 unmarried and recently married men from the 2011-2015 ACS data drawn
        randomly from the states of New York and California. The data frame contains the following variables:
       \describe{
         \item{state}{State of residence. This can be used to select appropriate cluster to sample alternate spouses}
         \item{perwt}{Person weight for the given respondent}
         \item{idh}{unique id of husband}
         \item{ageh}{age of husband}
         \item{raceh}{race of husband. Either "W" (white) or "B" (black)}
       }}
\source{
\url{https://usa.ipums.org/usa/}
}
\usage{
acs.malealters
}
\description{
A dataset containing information on a sample of unmarried and recently married men from the American Community
urvey 2011-2015 data for the states of New York and California. Data were downloaded from IPUMS.
}
\keyword{datasets}
