% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakeunion_functions.R
\name{generateCouples}
\alias{generateCouples}
\title{Generate Counterfactual Couples}
\usage{
generateCouples(n, geo, men, women, actual, id = "id", weight = NULL,
  keep = NULL, verbose = TRUE)
}
\arguments{
\item{n}{the number of counterfactual unions to create for each real union}

\item{geo}{a character string giving the name of the variable that identifies the clusters that alternate partners should be
sampled from. This variable must be named the same way in \code{men}, \code{women}, and \code{actual}.}

\item{men}{a data.frame object identifying all the potential "male" partners. This data.frame must include the 
variable identified by \code{geo} above and it must also contain an id variable whose name starts with the string given by
\code{id} and end with "h". Any person-specific variable that should be kept in the results should end with
an "h" (e.g. ageh, raceh). These variables should correspond exactly to variables in \code{actual} and should 
correspond to variables in \code{women} that end in "w".}

\item{women}{a data.frame object identifying all the potential "female" partners. Its format should be the same as \code{men}, except
variables should end in "w" rather than "h".}

\item{actual}{a data.frame identifying identifying the actual couples. This dataset should have the variable identified by \code{geo}
and variables for husbands and wives ids that end in "h" and "w" respectively. It should also contain all the same
        variables ending in "h" and "w" that should be kept in the final analysis.}

\item{weight}{A character string that identifies a weight variable for sampling of potential partners. If left \code{NULL}, partners will
be sampled with equal probability. This weight variable must exist in both the \code{men} and \code{women} data.}

\item{keep}{a vector of character strings identifying additional variables in \code{actual} that should be kept in the results.}

\item{verbose}{if true, the function will report its progress.}
}
\value{
The output of this program is a data.frame of actual and counterfactual unions. It will keep all variables in the three 
datasets that end in an "h" or "w" as well as:
\item{geo}{description the cluster identifier used in the function.}
\item{group}{a unique identifier based on the id of the spouse from the actual union who had partners sampled for them. This should be used as the
fixed effect in fixed effects models.}
\item{choice}{a boolean variables that is \code{TRUE} if this is an actual union and \code{FALSE} if this is a counterfactual union. This variable
should be used as the dependent variable in a fixed effect conditional logit model.}
}
\description{
This function will sample a set of alternate partners from a data.frame for a set of actual unions in order to
generate a set of counterfactual unions that can be used in a conditional logit model to predict the characteristics
that are likely to lead to a union match.
}
\details{
This program uses the \pkg{wrswoR} package to do fast weighted sampling without replacement. Large sets of alternate spouses and
actual unions can be sampled relatively quickly, but processing time will begin to increase exponentially with very large datasets
within each cluster.

For each actual union, the program randomly determines one of the two spouses to sample with 50/50 odds. This is to ensure that all characteristics
of a single spouse are fixed within the fixed effects conditional logit model. 

Currently the format of the datasets must be followed exactly for the function to work correctly.
}
