% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakeunion_functions.r
\name{organizeColumns}
\alias{organizeColumns}
\title{Organize columns for final union data}
\usage{
organizeColumns(couples, geo, keep = NULL)
}
\arguments{
\item{couples}{A data frame of the actual and counterfactual couples.}

\item{geo}{A character string giving the name of the clustering variable.}

\item{keep}{A vector of character strings giving additional variables to keep that don't end in
"h" or "w".}
}
\description{
This is an internal function used by \code{generateCouples} to organize the final
           dataset returned from the function.
}
