% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakeunion_functions.r
\name{poolChoiceModel}
\alias{poolChoiceModel}
\title{Run fixed effects conditional logit model on multiple datasets and pool results}
\usage{
poolChoiceModel(formula, datasets, method = "exact")
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}} specifying the \code{\link[survival]{clogit}} model to be
performed on each dataset.}

\item{datasets}{a list of datasets where each dataset is produced from the \code{\link{generateCouples}} function.}

\item{method}{A character string indicating the estimation method to use in the \code{\link[survival]{clogit}} model.}
}
\value{
a list containing the following objects:
\item{coefficients}{a data.frame object with the following elements:}
\itemize{
    \item{b.pool: }{The average coefficient across datasets.}
    \item{se.pool: }{standard error that combined within and between variance}
    \item{z.pool: }{z-statistic from dividing b by se}
    \item{pvalue.pool: }{p-value for the hypothesis test that the coefficient is zero in the population}
    \item{within.var: }{The square of the mean standard error across datasets}
    \item{between.var: }{the variance of the coefficient across datasets}
    }
\item{deviance}{A vector of deviances for each model.}
\item{bic}{A vector of BIC statistic for each dataset relative to the null model.}
}
\description{
Given a list of multiple datasets of real and counterfactual unionts and a model formula, this function
             will calculate a discrete choice model for each dataset and pool the results, taking account of both
             within and between variance in calculating the standard error.
}
\details{
Because the dataset of real and counterfactual unions is created by sampling among all the possible alternate partners,
         the results of models will vary as a result of this sampling process. Therefore, it may be useful to generate multiple
         datasets and pool model results across these datasets in a manner identical to multiple imputation, where the standard
         errors of estimates are adjusted for the variance in coefficient estimates across datasets.

         This function is a convenience function that will perform this pooling and produce properly adjusted results.
         The reported coefficients from the model are given by taking the mean across all datasets. The reported variance \eqn{V}
         for each parameter is given by:
         \deqn{V=W+(1+1/m)B}
         Where \eqn{m} is the number of datasets, \eqn{W} is the within variance, estimated by the square of the mean standard
         error across datasets, and \eqn{B} is the between variance estimated by the variance of coefficient estimates across
         datasets.

         Models are estimated using the \code{\link[survival]{clogit}} function from the \pkg{survival} package. This package
         must be installed.
}
\examples{
markets <- replicate(5, generateCouples(3,acs.couples,
                           acs.malealters,acs.femalealters,
                           "state",weight="perwt",verbose=FALSE),
                     simplify=FALSE)

poolChoiceModel(choice~ageh+I(ageh^2)+I(ageh-agew)+I((ageh-agew)^2)+strata(group),
                markets)
}
