% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakeunion_functions.R
\name{samplePartners}
\alias{samplePartners}
\title{Sample partners}
\usage{
samplePartners(actual, eligibles, n, partner, weight = NULL, id = "id")
}
\arguments{
\item{actual}{a data.frame of actual unions identical in structure to that for \code{generateCouples}.}

\item{eligibles}{a data.frame of eligible partners to be sampled. This should
have the same structure as that for \code{generateCouples}.}

\item{partner}{a character sting of  either "h" or "w" indicating which type of partner should be sampled.}

\item{weight}{optional name of weights variable in eligibles that 
allows for different probabilities of sampling eligible partners. If left out
then sampling will be done with equal probability}

\item{id}{a character string indicating the first part of the variable name for the id to identify indidividuals
in actual and eligibles. Identical to the argument in generateCouples.}
}
\value{
a data.frame object similar to \code{actual} but containing only counterfactual unions.
}
\description{
This function will draw a set of counterfactual partners for a given set of real unions.
It is primarily intended for internal use by the \code{generateCouples} functiom.
}
\details{
This function uses the \pkg{wrswoR} package for fast weighted sampling without replacement. In addition,
     for speed, the initial matching is done without concern for potentially sampling the correct partner from
     the set of alternate partners. This is then checked and removed in a secondary stage for the rare cases where
     such duplication occurs.
}
