% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakeunion_functions.r
\name{organizeColumns}
\alias{organizeColumns}
\title{Organize columns for final union data}
\usage{
organizeColumns(couples, geo, keep = NULL)
}
\description{
This is an internal function used by \code{generateCouples} to organize the final
           dataset returned from the function.
}
